ml-service/requirements.txt
kafka-python
torch
transformers
requests